LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/animation/column-wrap-reset-interpolation-expected.txt


FAIL Animate and reset column-wrap assert_equals: expected (string) "nowrap" but got (undefined) undefined
